+++ src/js/@shared/views/index.js
@@ -1,0 +1,1 @@
+export { linkButton } from './templates/linkButton.js';
