+++ src/js/userManage/components/UserAuth.js
@@ -1,17 +1,17 @@
-import { linkButton } from '../../@shared/views/linkButton';
-import { $ } from '../../@shared/utils/dom.js';
+import { linkButton } from '../../@shared/views';
+import { $ } from '../../@shared/utils';
 export class UserAuth {
   constructor(props) {
     this.props = props;
     this.selectDOM();
-    this.initRender();
+    this.render();
   }
 
   selectDOM() {
     this.$container = $('.sign-button-container');
   }
 
-  initRender() {
+  render() {
     this.$container.innerHTML = this.props.isSigned
       ? linkButton({ link: '#', text: '로그아웃' })
       : linkButton({ link: '#signin', text: '로그인' });
