+++ src/js/userManage/index.js
@@ -3,10 +3,10 @@ import { UserAuth } from './components/UserAuth';
 export class UserManage {
   constructor(props) {
     this.props = props;
-    this.mount();
+    this.moutChildComponents();
   }
 
-  mount() {
+  moutChildComponents() {
     this.userAuth = new UserAuth(this.props);
   }
 }
